aider: Hide the file extension in the list of results by stripping it before displaying the image name.
# Aider chat conversation:

USER: Please make it so that the file extension does not show in the list of results. We'll still need it obviously for displaying the images, but make it so it doesn't show up.

ASSISTANT: To hide the file extension in the list of results, we can modify the rendering logic to strip the extension before displaying the image name. We will still use the full name (with extension) for fetching the image.

Here are the changes:

src/index.tsx
```python
<<<<<<< SEARCH
                {image}
=======
                {image.replace(/\.[^/.]+$/, "")}
>>>>>>> REPLACE
```

This change will ensure that the file extension is not shown in the list of results.